\name{dtd_reliance}
\docType{data}
\alias{dtd_reliance}
\title{Daily data on market capitalization, debt and equity volatility for
  Reliance} 
\description{
  Sample dataset on the market capitalization (E), equity volatility
  (sE) and debt (F) of Reliance Industries Ltd. (a company listed on the
  National Stock Exchange, India). The data is at a daily frequency
  extending from Jan 4, 2011 to October 22, 2012.

  Equity volatility is the historical volatility of Reliance stock price
  computed on the basis of 250-day rolling window. Debt is calculated as
  the sum of all short term liabilities (one year and less) plus half
  the book value of all long term debt outstanding. The data is drawn
  from the CMIE's database, Prowess.
  
}

\usage{data(dtd_reliance)}

\format{data.frame (451 x 3)}
\keyword{datasets}
