% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ic}
\alias{ic}
\title{Function to compute impact cost.}
\usage{
ic(mbp, Q, d, partial = FALSE)
}
\arguments{
\item{mbp}{a data.frame, containing best price/quantity
pairs for buy and sell side with time stamps as row names.
Units:
\itemize{
\item Price (bbp/bsp): Some currency (INR, USD, etc.)
\item Quantities (bbq/bsq): Number of shares
}}

\item{Q}{an 'integer' vector containing the quantities
for which to compute the Impact Cost.}

\item{d}{Date}

\item{partial}{a 'logical' scalar, telling whether to do
partial executions}
}
\value{
A 'list' of matrices containing buy and sell IC.
'names' of the list are the Q values used. Each matrix will
contain rows having
\itemize{
\item buyIC percentage: numeric
\item sellIC percentage: numeric
}

The function returns \sQuote{NA} values if \dQuote{partial} is
 \sQuote{FALSE}, or, if there were no orders (empty book) at a
particular time.For the buy and sell ic measures the
corresponding row name for that entry is its timestamp.
}
\description{
Function to compute impact cost for a security for a day.
}
\examples{
## Demonstration of IC functions using the market by price data
## of CINEMAX from the ic_CINEMAX dataset.
try(data(package="ifrogs", "ic_CINEMAX"))
mbp <- cbind(as.data.frame(CINEMAX[1]), 
             as.data.frame(CINEMAX[2]), 
             as.data.frame(CINEMAX[3]), 
             as.data.frame(CINEMAX[4]))
print(mbp)

ans1 <- ic( mbp, Q = 600, d, partial = FALSE)
head(ans1)

ans2 <- ic( mbp, Q = 600, d, partial = TRUE)
head(ans2)
}
\author{
Chirag Anand
}
