\name{ic_CINEMAX}
\docType{data}
\alias{CINEMAX}
\title{Limit order book data of Cinemax India Ltd. for 29th October,
  2013}
\description{Twenty deep limit order book data consisting of the best
  buy and sell side prices and quantities for Cinemax India Ltd. (a
  company listed on the National Stock Exchange, India) for 29th
  October, 2013. The dataset is a list of four dataframe: buyQty,
  buyPrice, sellQty and sellPrice.
  
  - buyQty:	twenty levels of best buy quantities available for the day
  - buyPrice:	twenty levels of best sell prices available for the day
  - sellQty:	twenty levels of best sell quantities available for the day
  - sellPrice:	twenty levels of best sell prices available for the day
}

\usage{data(ic_CINEMAX)}

\format{list (4)}
\keyword{limit order book}
