\name{is_reliance}
\docType{data}
\alias{is_reliance}
\title{Example dataset on the spot and futures prices of RELIANCE} 

\description{
  Sample dataset for replication of results in the paper ``When do
  stock futures dominate price discovery?''. The dataset contains spot
  and futures prices of RELIANCE (traded on NSE, India) for a day of
  data used in the study. The prices are at one second frequency for 6
  May, 2009. 
  
  The first column represents the date and time while the second and the
  third column represent the stock price on the spot and the futures
  market respectively. 
}

\references{
  Aggarwal N, Thomas S (2013), When do stock futures dominate price
  discovery, IGIDR
  WP-2011-016. \url{http://www.igidr.ac.in/pdf/publication/WP-2011-016.pdf}  
}
  
\usage{data(is_reliance)}
\format{data.frame (20094 x 3)}
\keyword{datasets}
