\name{lob}
\docType{data}
\alias{lob}
\title{A snapshot of the RELIANCE spot limit order book at 11AM on 1st January,
2009.}
\description{
  The data is a snapshot of the limit order book for a stock
  RELIANCE trading on the National Stock Exchange of India on 1st
  January 2009. The snapshot has been taken at 11AM. The data consists
  of pairs of limit order price-limit order quantity present in the
  limit order book at the time.
  
  The snaphot pertains to the ask side of the limit order book and
  hence, the limit prices are arranged in increasing order of price priority.
}
\usage{data(lob)}
\format{data.frame (8561 x 2)}
\keyword{datasets}
