\name{vix_nifty}

\docType{data}

\alias{vix_nifty}

\title{A cross section of Nifty options on 1st September, 2010.}

\description{
  A sample of Nifty options at NSE from 14:59:45 to 15:00:00 on 1st September,
  2010. This includes the two nearest maturities, one and three month riskfree
  rates, option type, strike price, price of the underlying index, and the best
  buy and ask prices of the options.
}

\usage{data(vix_nifty)}

\format{list of 2 data.frames}

\keyword{datasets}
