\name{vxo_nifty}

\docType{data}

\alias{vxo_nifty}

\title{A cross section of Nifty options on 1st September, 2010.}

\description{
  A sample of eight at-the-money (ATM) Nifty options at NSE from 14:59:45 to
  15:00:00 on 1st September, 2010. This includes the two nearest maturities,
  one and three month riskfree rates, option type, strike price, price of the
  underlying index, and the best buy and ask prices of the options.
}

\usage{data(vxo_nifty)}

\format{data.frame (8 x 7)}

\keyword{datasets}
