\name{vxo_spx}

\docType{data}

\alias{vxo_spx}

\title{A cross section of end-of-day SPX options on 17th September, 2010.}

\description{
  The end-of-day data on eight at-the-money (ATM) S&P 500 (SPX) index options
  at CBOE for 17th September, 2010. This includes the two nearest maturities,
  one and three month riskfree rates, option type, strike price, price of the
  underlying index, and the best buy and ask prices of the options.
}

\usage{data(vxo_spx)}

\format{data.frame (8 x 7)}

\keyword{datasets}
